+++ packages/dart/test/date_test.dart
@@ -3,6 +3,11 @@ import 'package:test/test.dart';
 
 void main() {
   group('Date', () {
+    test('dayDiff', () {
+      expect(dayDiff(DateTime(2021, 1, 1), DateTime(2021, 1, 2)), 1);
+      expect(dayDiff(DateTime(2021, 1, 1), DateTime(2021, 2, 28)), 58);
+    });
+
     test('isValidDate', () {
       expect(isValidDate('2021-01-01'), true);
       expect(isValidDate('2021-02-28'), true);